requirements.txt
# Core FastAPI and Web Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.19

# Database - SQLAlchemy and Drivers
sqlalchemy==2.0.36
alembic==1.14.0

# For SQLite (development) - built into Python, no package needed

# For PostgreSQL on Azure (commented out for now - we'll install when needed on Day 2)
# psycopg2-binary==2.9.10

# Pydantic for validation
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# HTTP Requests (for API integrations)
httpx==0.28.1
requests==2.32.3

# Azure SDK
azure-identity==1.19.0
azure-storage-blob==12.24.0
azure-keyvault-secrets==4.9.0

# Date/Time handling
python-dateutil==2.9.0.post0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0

# Code Quality
black==24.10.0
flake8==7.1.1
mypy==1.13.0

# Documentation
mkdocs==1.6.1
mkdocs-material==9.5.48